requirements.txt
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.2
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pyaes==1.6.1
pyasn1==0.5.1
python-dotenv==1.0.1
python-telegram-bot==20.8
requests==2.31.0
rsa==4.9
sniffio==1.3.1
Telethon==1.34.0
urllib3==2.2.1
Werkzeug==3.0.1
